% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_values.R
\name{download}
\alias{download}
\title{Download model objects}
\usage{
download(file, url)
}
\arguments{
\item{file}{A string that provides the file name, for example "rb_mod1a.rda"}

\item{url}{The location of the file, for example "https://osf.io/eq9rw/download"}
}
\description{
Download model objects or variable lists
}
\examples{
#load package
library(writeAlizer)

#download rb_mod1a to extdata
download("rb_mod1a.rda", "https://osf.io/eq9rw/download")
}
\author{
Sterett H. Mercer <sterett.mercer@ubc.ca>
}
