% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_values.R
\name{download_mod}
\alias{download_mod}
\title{Download predictive models}
\usage{
download_mod(model)
}
\arguments{
\item{model}{A string telling which scoring model to use.
Options are:
'rb_mod1', 'rb_mod2', 'rb_mod3narr', 'rb_mod3exp',
'rb_mod3per', 'rb_mod3all', 'rb_mod3narr_v2', 'rb_mod3exp_v2',
'rb_mod3per_v2', or 'rb_mod3all_v2' for ReaderBench files,
'coh_mod1', 'coh_mod2', 'coh_mod3narr', 'coh_mod3exp', 'coh_mod3per'
 or 'coh_mod3all' for Coh-Metrix files, or
'gamet_cws1' for GAMET files}
}
\description{
Download models so that predicted quality scores can be generated
in the predict_quality function
}
\examples{
#load package
library(writeAlizer)

#Download ReaderBench Model 2
download_mod('rb_mod2')
}
\seealso{
\code{\link{predict_quality}}
}
\author{
Sterett H. Mercer <sterett.mercer@ubc.ca>
}
