% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{download_model}
\alias{download_model}
\title{Download scoring models to the working directory.}
\usage{
download_model(mod)
}
\arguments{
\item{mod}{A string telling which model group to download. Options are
'rb_mod1' for ReaderBench files to generate holistic quality,
'coh_mod1' for Coh-Metrix files to generate holistic quality,
and 'rb_gamet_cws1' for a merged ReaderBench and GAMET file
to generate Correct Word Sequences (CWS) and Correct Minus Incorrect
Word Sequences (CIWS) scores}
}
\description{
Download scoring models to the working directory.
}
\examples{
#To download models to process ReaderBench files:
\dontrun{
download_model('rb_mod1')
}
#To download models to process Coh-Metrix files:
\dontrun{
download_model('coh_mod1')
}
#To download models to generate CWS/CIWS scores from ReaderBench and GAMET files:
\dontrun{
download_model('rb_gamet_cws1')
}
}
