% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{import_coh}
\alias{import_coh}
\title{Import a Coh-Metrix output file (.csv) into R.}
\usage{
import_coh(path)
}
\arguments{
\item{path}{A string giving the path and filename to import.}
}
\description{
Import a Coh-Metrix output file (.csv) into R.
}
\seealso{
\code{\link{predict_quality}}
}
