% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{import_coh}
\alias{import_coh}
\title{Import a Coh-Metrix output file(.csv) into R.}
\usage{
import_coh(path)
}
\arguments{
\item{path}{A string giving the path and filename to import.}
}
\description{
Import a Coh-Metrix output file(.csv) into R.
}
\examples{
#To import as 'coh_file' a Coh-Metrix file (sample name: coh_output.csv)
#that is stored in the working directory
\dontrun{
coh_file <- import_coh('coh_output.csv')
}
}
\author{
Sterett H. Mercer <sterett.mercer@ubc.ca> and Xuejun Ji
}
