% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{import_gamet}
\alias{import_gamet}
\title{Import a GAMET output file into R.}
\usage{
import_gamet(path)
}
\arguments{
\item{path}{A string giving the path and filename to import.}
}
\description{
Import a GAMET output file into R.
}
\examples{
#To import as 'gamet_file' a GAMET file (sample name: gamet_output.csv)
# that is stored in the working directory
\dontrun{
gamet_file <- import_gamet('gamet_output.csv')
}
}
\author{
Sterett H. Mercer <sterett.mercer@ubc.ca> and Xuejun Ji
}
