% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{import_gamet}
\alias{import_gamet}
\title{Import a GAMET output file into R.}
\usage{
import_gamet(path)
}
\arguments{
\item{path}{A string giving the path and filename to import.}
}
\description{
Import a GAMET output file into R.
}
\seealso{
\code{\link{predict_quality}}
}
