% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{import_merge_gamet_rb}
\alias{import_merge_gamet_rb}
\title{Import a ReaderBench output file(.xlsx) and GAMET output file (.csv) into R, and merge the two files.
ReaderBench output files (.csv) should be converted to Excel format (.xlsx) before import.}
\usage{
import_merge_gamet_rb(rb_path, gamet_path)
}
\arguments{
\item{rb_path}{A string giving the path and ReaderBench filename to import.}

\item{gamet_path}{A string giving the path and GAMET filename to import.}
}
\description{
Import a ReaderBench output file(.xlsx) and GAMET output file (.csv) into R, and merge the two files.
ReaderBench output files (.csv) should be converted to Excel format (.xlsx) before import.
}
\examples{
#To import as 'rb_gam_file' a ReaderBench file (sample name: rb_output.xlsx)
#and GAMET file (sample name: gamet_output.csv) stored in the working
#directory and then merge them
\dontrun{
rb_gam_file <- import_merge_gamet_rb('rb_output.xlsx', 'gamet_output.csv')
}
}
\author{
Sterett H. Mercer <sterett.mercer@ubc.ca> and Xuejun Ji
}
