% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{import_merge_gamet_rb}
\alias{import_merge_gamet_rb}
\title{Import a ReaderBench output file (.csv) and GAMET output file (.csv),
and merge the two files on ID.}
\usage{
import_merge_gamet_rb(rb_path, gamet_path)
}
\arguments{
\item{rb_path}{A string giving the path and ReaderBench filename to import.}

\item{gamet_path}{A string giving the path and GAMET filename to import.}
}
\description{
Import a ReaderBench output file (.csv) and GAMET output file (.csv),
and merge the two files on ID.
}
\examples{
# Example with package sample data
rb_path   <- system.file("extdata", "sample_rb.csv", package = "writeAlizer")
gam_path  <- system.file("extdata", "sample_gamet.csv", package = "writeAlizer")
rb_gam    <- import_merge_gamet_rb(rb_path, gam_path)
head(rb_gam)
}
\seealso{
\code{\link{predict_quality}}
}
