% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{import_rb}
\alias{import_rb}
\title{Import a ReaderBench output file (.csv) into R.}
\usage{
import_rb(path)
}
\arguments{
\item{path}{A string giving the path and filename to import.}
}
\description{
Import a ReaderBench output file (.csv) into R.
}
\seealso{
\code{\link{predict_quality}}
}
