% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{import_rb}
\alias{import_rb}
\title{Import a ReaderBench output file(.xlsx) into R. ReaderBench output files (.csv)
should be converted to Excel format (.xlsx) before import.}
\usage{
import_rb(path)
}
\arguments{
\item{path}{A string giving the path and filename to import.}
}
\description{
Import a ReaderBench output file(.xlsx) into R. ReaderBench output files (.csv)
should be converted to Excel format (.xlsx) before import.
}
\examples{
#To import as 'rb_file' a ReaderBench file (sample name: rb_output.xlsx)
#that is stored in the working directory
\dontrun{
rb_file <- import_rb('rb_output.xlsx')
}
}
\author{
Sterett H. Mercer <sterett.mercer@ubc.ca> and Xuejun Ji
}
