% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{import_rb}
\alias{import_rb}
\title{Import a ReaderBench output file(.csv) into R.}
\usage{
import_rb(x)
}
\arguments{
\item{path}{A string giving the path and filename to import.}
}
\description{
Import a ReaderBench output file(.csv) into R.
}
\examples{
#' ##Example 1:
#Using a sample data file included with writeAlizer package

#load package
library(writeAlizer)

#get path of sample ReaderBench output file
file_path <- system.file("extdata", "sample_rb.csv", package = "writeAlizer")

#see path to sample file
file_path

#import file and store as "rb_file"
rb_file <- import_rb(file_path)

##Example 2:
#To import as "rb_file" a ReaderBench file (sample name: rb_output.csv)
#that is stored in the working directory
\dontrun{
rb_file <- import_rb("rb_output.csv")
}

}
\seealso{
\code{\link{predict_quality}}
}
