% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{import_rb}
\alias{import_rb}
\title{Import a ReaderBench output file(.xlsx) into R. ReaderBench output files (.csv)
should be converted to Excel format (.xlsx) before import.}
\usage{
import_rb(path)
}
\arguments{
\item{path}{A string giving the path and filename to import.}
}
\description{
Import a ReaderBench output file(.xlsx) into R. ReaderBench output files (.csv)
should be converted to Excel format (.xlsx) before import.
}
\examples{
#' ##Example 1:
#Using a sample data file included with writeAlizer package

#load package
library(writeAlizer)

#get path of sample ReaderBench output file
file_path <- system.file("extdata", "sample_rb.xlsx", package = "writeAlizer")

#see path to sample file
file_path

#import file and store as "rb_file"
rb_file <- import_rb(file_path)

##Example 2:
#To import as "rb_file" a ReaderBench file (sample name: rb_output.xlsx)
#that is stored in the working directory
\dontrun{
rb_file <- import_rb("rb_output.xlsx")
}
}
\seealso{
\code{\link{predict_quality}}
}
\author{
Sterett H. Mercer <sterett.mercer@ubc.ca> and Xuejun Ji
}
