% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_values.R
\name{predict_quality}
\alias{predict_quality}
\title{Apply scoring models for predictions}
\usage{
predict_quality(model, data, store = FALSE, name = "filename.csv")
}
\arguments{
\item{model}{A string telling which scoring model to use.
Options are:
'rb_mod1' for ReaderBench files to generate holistic quality,
'coh_mod1' for Coh-Metrix files to generate holistic quality,
and 'rb_gamet_cws1' for a merged ReaderBench and GAMET file
to generate Correct Word Sequences (CWS) and Correct Minus Incorrect
Word Sequences (CIWS) scores}

\item{data}{The name of the R object corresponding to the data file. The
import_gamet(), import_coh(), import_rb(), and import_merge_gamet_rb()
functions should be used before this function
to generate these data objects.}

\item{store}{When store = TRUE, this function will generate scores, merge
the scores into the data file, and export the file as .csv in the
working directory. When store = FALSE (the default) the predicted scores
are returned to a user-specified object or the R console.}

\item{name}{When store = TRUE, the name parameter gives the filename for
the .csv file (for example, "filename.csv") that is
generated to the working directory.}
}
\description{
Apply scoring models to ReaderBench, CohMetrix, and/or
GAMET files. Holistic writing quality can be
generated from Readerbench (model = 'rb_mod1') or
Coh-Metrix files (model = 'coh_mod1'). Also,
Correct Word Sequences and Correct Minus Incorrect
Word Sequences can be generated from a combined
Readerbench and GAMET file.
}
\examples{
#To generate holistic quality from a ReaderBench file ('rb_data')
#and return scores to an object called 'rb_quality':
\dontrun{
rb_quality <- predict_quality('rb_mod1', rb_data, store = FALSE)
}
#To generate holistic quality from a Coh-Metrix file ('coh_data'), merge
#predicted scores with the data file, and store as "output.csv" in the
#working directory
\dontrun{
predict_quality('coh_mod1', coh_data, store = TRUE, name = "output.csv")
}
}
\author{
Sterett H. Mercer <sterett.mercer@ubc.ca>
}
