% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_values.R
\name{preprocess}
\alias{preprocess}
\title{Pre-process data}
\usage{
preprocess(model, data)
}
\arguments{
\item{model}{A string telling which scoring model to use.
Options are:
'rb_mod1', 'rb_mod2', 'rb_mod3narr', 'rb_mod3exp',
'rb_mod3per', 'rb_mod3all', 'rb_mod3narr_v2', 'rb_mod3exp_v2',
'rb_mod3per_v2', or 'rb_mod3all_v2' for ReaderBench files,
'coh_mod1', 'coh_mod2', 'coh_mod3narr', 'coh_mod3exp', 'coh_mod3per'
 or 'coh_mod3all' for Coh-Metrix files, or
'gamet_cws1' for GAMET files}

\item{data}{The name of the R object corresponding to the data file. Use
\code{\link{import_gamet}}, \code{\link{import_coh}}, or \code{\link{import_rb}}
before this function to generate these data objects.}
}
\value{
A list of pre-processed data frames, one per sub-model
}
\description{
Pre-process Coh-Metrix and ReaderBench data files before applying predictive models
}
\author{
Sterett H. Mercer <sterett.mercer@ubc.ca>
}
