% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{wa_download}
\alias{wa_download}
\alias{download}
\title{Download artifact into cache with optional checksum}
\usage{
wa_download(file, url, sha256 = NULL, quiet = TRUE)

download(file, url) # deprecated
}
\description{
Download artifact into cache with optional checksum
}
