% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAlizer-package.R
\docType{package}
\name{writeAlizer-package}
\alias{writeAlizer-package}
\title{writeAlizer: Generate Predicted Writing Quality Scores}
\description{
Imports variables from 'ReaderBench' (Dascalu et al., 2018)\doi{10.1007/978-3-319-66610-5_48}, 'Coh-Metrix' (McNamara et al., 2014)\doi{10.1017/CBO9780511894664}, and/or 'GAMET' (Crossley et al., 2019) \doi{10.17239/jowr-2019.11.02.01} output files; downloads predictive scoring models described in Mercer & Cannon (2022)\doi{10.31244/jero.2022.01.03} and Mercer et al.(2021)\doi{10.1177/0829573520987753}; and generates predicted writing quality and curriculum-based measurement (McMaster & Espin, 2007)\doi{10.1177/00224669070410020301} scores.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shmercer/writeAlizer/}
  \item Report bugs at \url{https://github.com/shmercer/writeAlizer/issues}
}

}
\author{
\strong{Maintainer}: Sterett H. Mercer \email{sterett.mercer@ubc.ca} (\href{https://orcid.org/0000-0002-7940-4221}{ORCID})

}
\keyword{internal}
