% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAlizer-package.R
\docType{package}
\name{writeAlizer-package}
\alias{writeAlizer-package}
\title{writeAlizer: Generate Predicted Writing Quality and Written Expression CBM Scores}
\description{
Imports ReaderBench, Coh-Metrix, and GAMET output files; downloads predictive scoring models; and computes predicted writing quality scores from these inputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shmercer/writeAlizer/}
  \item Report bugs at \url{https://github.com/shmercer/writeAlizer/issues}
}

}
\author{
\strong{Maintainer}: Sterett H. Mercer \email{sterett.mercer@ubc.ca} (\href{https://orcid.org/0000-0002-7940-4221}{ORCID})

}
\keyword{internal}
