% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAlizer-package.R
\docType{package}
\name{writeAlizer-package}
\alias{writeAlizer-package}
\title{writeAlizer: Generate Predicted Writing Quality and Written Expression CBM Scores}
\description{
This package (a) imports ReaderBench, Coh-Metrix, and GAMET output files into R, (b) downloads existing predictive scoring models to the local machine, and (c) uses the predictive scoring models to generate predicted writing quality scores or Correct Word Sequences and Correct Minus Incorrect Word Sequences scores from the ReaderBench, Coh-Metrix, and/or GAMET files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shmercer/writeAlizer/}
  \item Report bugs at \url{https://github.com/shmercer/writeAlizer/issues}
}

}
\author{
\strong{Maintainer}: Mercer Sterett H. \email{sterett.mercer@ubc.ca}

}
\keyword{internal}
